Extract HIR match lowering helpers
Move enum and const match lowering into a compiler.hir child module and lower the compiler source monolith ratchet for the reduced HIR facade.

Validation: cargo fmt --manifest-path stage1/crates/axiomc/Cargo.toml; cargo test --manifest-path stage1/Cargo.toml -p axiomc --lib hir::boundary_tests -- --nocapture; make stage1-hir-boundary; make stage1-compiler-source-monoliths-test; make stage1-compiler-source-monoliths; cargo test --manifest-path stage1/Cargo.toml -p axiomc --lib; git diff --check; git diff --cached --check
